+++ typings/events/ready.d.ts
@@ -1,2 +1,0 @@
-import type { ShewenyClient } from '../client/Client';
-export default function run(client: ShewenyClient): Promise<void>;