+++ typings/helpers/ShewenyError.d.ts
@@ -1,4 +1,0 @@
-import { ShewenyClient } from '../client/Client';
-export declare class ShewenyError extends Error {
-    constructor(client: ShewenyClient, err: string | Error, ...args: string[]);
-}