+++ typings/helpers/ShewenyInformation.d.ts
@@ -1,9 +1,0 @@
-import type { ShewenyClient } from '../client/Client';
-/**
- * Information class for Sheweny.
- */
-export declare class ShewenyInformation {
-    client: ShewenyClient;
-    message: string;
-    constructor(client: ShewenyClient, message: string);
-}