+++ typings/managers/index.d.ts
@@ -1,9 +1,0 @@
-/**
- * Sheweny Managers: Classes to manage the different structures (registrations, loading, etc.)
- */
-export { BaseManager } from './BaseManager';
-export { ButtonsManager } from './ButtonsManager';
-export { CommandsManager } from './CommandsManager';
-export { EventsManager } from './EventsManager';
-export { InhibitorsManager } from './InhibitorsManager';
-export { SelectMenusManager } from './SelectMenusManager';