+++ typings/utils/client-util.d.ts
@@ -1,4 +1,0 @@
-import { Intents } from 'discord.js';
-import type { PartialTypes } from 'discord.js';
-export declare function getAllIntents(): Intents;
-export declare function getAllPartials(): PartialTypes[];